// iot-service/iot-angular/src/sdk/core/models/send-relay-command-dto.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface SendRelayCommandDto {

  /**
   * Action to perform
   */
  action: 'on' | 'off' | 'pulse';

  /**
   * Device ID (MAC address)
   */
  deviceId: string;

  /**
   * Duration in milliseconds (only for PULSE action)
   */
  duration?: number;

  /**
   * Target relay output
   */
  target: 'out1' | 'out2';
}