// iot-service/iot-angular/src/sdk/core/models/command-response-dto.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface CommandResponseDto {

  /**
   * Response message
   */
  message: string;

  /**
   * Command payload sent
   */
  payload: {
};

  /**
   * Success status
   */
  success: boolean;

  /**
   * Timestamp when command was sent
   */
  timestamp: string;

  /**
   * MQTT topic used
   */
  topic: string;
}